fix UI, logic Order

+++ src/main/java/com/webbanhang/webbanhang/DTO/request/Product/ProductRequestDTO.java
@@ -1,14 +1,8 @@
 package com.webbanhang.webbanhang.DTO.request.Product;
 
-import com.webbanhang.webbanhang.Util.Validation.PhoneNumber;
-import jakarta.persistence.Column;
-import jakarta.validation.constraints.Email;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
-import jakarta.validation.constraints.Pattern;
 import lombok.*;
-import org.springframework.format.annotation.NumberFormat;
-import org.springframework.web.multipart.MultipartFile;
 
 import java.io.Serializable;
 
@@ -43,13 +37,13 @@ public class ProductRequestDTO implements Serializable {
     private String description;
 
 
-    private Double unitPrice;
+    private Integer unitPrice;
 
 
     private Integer quantity;
 
 
-    private Double unitCost;
+    private Integer unitCost;
     @NotBlank(message = "supplierID must be not blank")
     @NotNull
     private String supplierID;
